localizer funcionando com o ouster! :fire:

+++ src/velodyne/velodyne_interface.cpp
@@ -207,7 +207,7 @@ carmen_velodyne_estimate_shot_time(double sensor_last_timestamp, double sensor_t
 void
 carmen_velodyne_partial_scan_update_points(carmen_velodyne_partial_scan_message *velodyne_message,
 		int vertical_resolution, spherical_point_cloud *points, unsigned char *intensity,
-		int *ray_order, double *vertical_correction, float range_max, double timestamp)
+		int *ray_order, double *vertical_correction, double range_max, double timestamp)
 {
 	points->timestamp = timestamp;
 
@@ -233,7 +233,7 @@ carmen_velodyne_partial_scan_update_points(carmen_velodyne_partial_scan_message 
 void
 carmen_velodyne_partial_scan_update_points_with_remission_check(carmen_velodyne_partial_scan_message *velodyne_message,
 		int vertical_resolution, spherical_point_cloud *points, unsigned char *intensity,
-		int *ray_order, double *vertical_correction, float range_max, double timestamp, int use_remission)
+		int *ray_order, double *vertical_correction, double range_max, double timestamp, int use_remission)
 {
 	points->timestamp = timestamp;
 
@@ -264,7 +264,7 @@ carmen_velodyne_partial_scan_update_points_with_remission_check(carmen_velodyne_
 
 void
 variable_scan_update_points_with_remission_check(carmen_velodyne_variable_scan_message *msg, int vertical_resolution, spherical_point_cloud *points,
-		unsigned char *intensity, int *ray_order, double *vertical_correction, float range_max, double range_division_factor, double timestamp, int use_remission)
+		unsigned char *intensity, int *ray_order, double *vertical_correction, double range_max, double range_division_factor, double timestamp, int use_remission)
 {
 	points->timestamp = timestamp;
 
@@ -292,7 +292,7 @@ variable_scan_update_points_with_remission_check(carmen_velodyne_variable_scan_m
 void
 carmen_velodyne_variable_scan_update_points(carmen_velodyne_variable_scan_message *message,
 		int vertical_resolution, spherical_point_cloud *points, unsigned char *intensity,
-		int *ray_order, double *vertical_correction, float range_max, double timestamp)
+		int *ray_order, double *vertical_correction, double range_max, double timestamp)
 {
 	points->timestamp = timestamp;
 
@@ -377,42 +377,43 @@ carmen_velodyne_free_variable_velodyne_message(
 }
 
 
- void
- load_lidar_config(int argc, char** argv, int lidar_id, carmen_lidar_config **lidar_config)
- {
-     char *vertical_angles_string, *ray_order_string;
-     char lidar_string[256];
+void
+load_lidar_config(int argc, char** argv, int lidar_id, carmen_lidar_config **lidar_config)
+{
+    char *vertical_angles_string, *ray_order_string;
+    char lidar_string[256];
 
  	carmen_lidar_config *lidar_config_p = *lidar_config;
 	
  	sprintf(lidar_string, "lidar%d", lidar_id);        // Geather the lidar id
 
-    carmen_param_t param_list[] = {
- 			{lidar_string, (char*)"model", CARMEN_PARAM_STRING, &lidar_config_p->model, 0, NULL},
- 			{lidar_string, (char*)"ip", CARMEN_PARAM_STRING, &lidar_config_p->ip, 0, NULL},
- 			{lidar_string, (char*)"port", CARMEN_PARAM_STRING, &lidar_config_p->port, 0, NULL},
- 			{lidar_string, (char*)"shot_size", CARMEN_PARAM_INT, &lidar_config_p->shot_size, 0, NULL},
-             {lidar_string, (char*)"min_sensing", CARMEN_PARAM_INT, &lidar_config_p->min_sensing, 1, NULL},
-             {lidar_string, (char*)"max_sensing", CARMEN_PARAM_INT, &lidar_config_p->max_sensing, 1, NULL},
- 			{lidar_string, (char*)"range_division_factor", CARMEN_PARAM_INT, &lidar_config_p->range_division_factor, 0, NULL},
- 			{lidar_string, (char*)"max_range", CARMEN_PARAM_DOUBLE, &lidar_config_p->max_range, 1, NULL},
-             {lidar_string, (char*)"time_between_shots", CARMEN_PARAM_DOUBLE, &lidar_config_p->time_between_shots, 0, NULL},
- 			{lidar_string, (char*)"x", CARMEN_PARAM_DOUBLE, &(lidar_config_p->pose.position.x), 1, NULL},
- 			{lidar_string, (char*)"y", CARMEN_PARAM_DOUBLE, &(lidar_config_p->pose.position.y), 1, NULL},
- 			{lidar_string, (char*)"z", CARMEN_PARAM_DOUBLE, &lidar_config_p->pose.position.z, 1, NULL},
- 			{lidar_string, (char*)"roll", CARMEN_PARAM_DOUBLE, &lidar_config_p->pose.orientation.roll, 1, NULL},
- 			{lidar_string, (char*)"pitch", CARMEN_PARAM_DOUBLE, &lidar_config_p->pose.orientation.pitch, 1, NULL},
- 			{lidar_string, (char*)"yaw", CARMEN_PARAM_DOUBLE, &lidar_config_p->pose.orientation.yaw, 1, NULL},
- 			{lidar_string, (char*)"ray_order", CARMEN_PARAM_STRING, &ray_order_string, 0, NULL},
- 			{lidar_string, (char*)"vertical_angles", CARMEN_PARAM_STRING, &vertical_angles_string, 0, NULL},
+    carmen_param_t param_list[] =
+    {
+		{lidar_string, (char *) "model", CARMEN_PARAM_STRING, &lidar_config_p->model, 0, NULL},
+		{lidar_string, (char *) "ip", CARMEN_PARAM_STRING, &lidar_config_p->ip, 0, NULL},
+		{lidar_string, (char *) "port", CARMEN_PARAM_STRING, &lidar_config_p->port, 0, NULL},
+		{lidar_string, (char *) "shot_size", CARMEN_PARAM_INT, &lidar_config_p->shot_size, 0, NULL},
+		{lidar_string, (char *) "min_sensing", CARMEN_PARAM_INT, &lidar_config_p->min_sensing, 1, NULL},
+		{lidar_string, (char *) "max_sensing", CARMEN_PARAM_INT, &lidar_config_p->max_sensing, 1, NULL},
+		{lidar_string, (char *) "range_division_factor", CARMEN_PARAM_INT, &lidar_config_p->range_division_factor, 0, NULL},
+		{lidar_string, (char *) "max_range", CARMEN_PARAM_DOUBLE, &lidar_config_p->max_range, 1, NULL},
+		{lidar_string, (char *) "time_between_shots", CARMEN_PARAM_DOUBLE, &lidar_config_p->time_between_shots, 0, NULL},
+		{lidar_string, (char *) "x", CARMEN_PARAM_DOUBLE, &(lidar_config_p->pose.position.x), 1, NULL},
+		{lidar_string, (char *) "y", CARMEN_PARAM_DOUBLE, &(lidar_config_p->pose.position.y), 1, NULL},
+		{lidar_string, (char *) "z", CARMEN_PARAM_DOUBLE, &lidar_config_p->pose.position.z, 1, NULL},
+		{lidar_string, (char *) "roll", CARMEN_PARAM_DOUBLE, &lidar_config_p->pose.orientation.roll, 1, NULL},
+		{lidar_string, (char *) "pitch", CARMEN_PARAM_DOUBLE, &lidar_config_p->pose.orientation.pitch, 1, NULL},
+		{lidar_string, (char *) "yaw", CARMEN_PARAM_DOUBLE, &lidar_config_p->pose.orientation.yaw, 1, NULL},
+		{lidar_string, (char *) "ray_order", CARMEN_PARAM_STRING, &ray_order_string, 0, NULL},
+		{lidar_string, (char *) "vertical_angles", CARMEN_PARAM_STRING, &vertical_angles_string, 0, NULL},
  	};
  	carmen_param_install_params(argc, argv, param_list, (sizeof(param_list) / sizeof(param_list[0])));
 
-    lidar_config_p->vertical_angles = (double*) malloc(lidar_config_p->shot_size * sizeof(double));
- 	lidar_config_p->ray_order = (int*) malloc(lidar_config_p->shot_size * sizeof(int));
+    lidar_config_p->vertical_angles = (double *) malloc(lidar_config_p->shot_size * sizeof(double));
+ 	lidar_config_p->ray_order = (int *) malloc(lidar_config_p->shot_size * sizeof(int));
  		
 	for (int i = 0; i < lidar_config_p->shot_size; i++)
- 		lidar_config_p->ray_order[i] = CLF_READ_INT(&ray_order_string); // CLF_READ_DOUBLE takes a double number from a string
+ 		lidar_config_p->ray_order[i] = CLF_READ_INT(&ray_order_string); // CLF_READ_INT takes an int number from a string
 
    	for (int i = 0; i < lidar_config_p->shot_size; i++)
  		lidar_config_p->vertical_angles[i] = CLF_READ_DOUBLE(&vertical_angles_string); // CLF_READ_DOUBLE takes a double number from a string
@@ -424,4 +425,4 @@ carmen_velodyne_free_variable_velodyne_message(
  	// 	printf("%lf ", lidar_config_p->vertical_angles[i]); printf("\n");
  	// for (int i = 0; i < lidar_config_p->shot_size; i++)
  	// 	printf("%d ", lidar_config_p->ray_order[i]); printf("\n");
- }
+}
